Refactor numerical calculations to ensure float precision across various modules
- Updated confidence calculations in anomaly detection, risk controllers, and strategy modules to explicitly cast to float for improved numerical stability.
- Enhanced performance metrics and scoring systems in voting and simulation modules by ensuring consistent float handling.
- Improved clarity and maintainability by removing redundant type casts and ensuring all relevant calculations are performed in float context.
- Added aiohttp dependency for async HTTP client support in visualization module.

+++ modules/meta/metar_rl_controller.py
@@ -103,11 +103,11 @@ class AgentPerformanceTracker:
                 recent_avg = np.mean(recent_rewards)
                 older_avg = np.mean(older_rewards)
                 improvement = (recent_avg - older_avg) / (abs(older_avg) + 1e-8)
-                metrics['convergence_score'] = max(0, min(1, improvement + 0.5))
+                metrics['convergence_score'] = max(0, min(1, float(improvement) + 0.5))
             
             # Calculate stability score
             reward_std = np.std(recent_rewards)
-            metrics['stability_score'] = max(0, 1 - reward_std / 100.0)
+            metrics['stability_score'] = max(0, float(1 - reward_std / 100.0))
     
     def get_best_agent(self) -> str:
         """Get the name of the best performing agent"""
